==> notebooks/00_data_split.ipynb <==
import pandas as pd

df = pd.read_csv('C:\\Users\\Mustapha\\Desktop\\ML-Project-Scratch\\data\\raw\\untouched_raw_original.csv')

pd.set_option('display.max_columns', None) # To display all columns
pd.set_option('display.max_rows', None) # To display all rows

df.shape

df.head(2)

df['date'] = pd.to_datetime(df['date'])
print(df['date'].dt.year.value_counts())

##### Avoid Leakage

- train set (2012–2019) is “clean” historical training.

- eval set (2020–2021) is “unseen” during training, but still used for validation and hyperparameter tuning.

- holdout set (2022–2023) is completely untouched until the very end.

# Sort by date (important to avoid shuffling issues)
df = df.sort_values("date")

cutoff_date_eval = "2020-01-01" # validation set starts from Jan 1, 2020
cutoff_date_holdout = "2022-01-01" # holdout set starts from Jan 1, 2022

# train set before 2020
train_df = df[df['date'] < cutoff_date_eval]

# evaluation/validation set between 2020 and 2021
eval_df = df[(df['date'] >= cutoff_date_eval) & (df['date'] < cutoff_date_holdout)]

# holdout/test set from 2022 onwards
holdout_df = df[df['date'] >= cutoff_date_holdout]

print(f"Training set shape: {train_df.shape}")
print(f"Evaluation set shape: {eval_df.shape}")
print(f"Holdout set shape: {holdout_df.shape}")

train_df.head(2)

eval_df.head(2)

holdout_df.head(2)

train_df.to_csv('C:\\Users\\Mustapha\\Desktop\\ML-Project-Scratch\\data\\raw\\train.csv', index=False)
eval_df.to_csv('C:\\Users\\Mustapha\\Desktop\\ML-Project-Scratch\\data\\raw\\eval.csv', index=False)
holdout_df.to_csv('C:\\Users\\Mustapha\\Desktop\\ML-Project-Scratch\\data\\raw\\holdout.csv', index=False)
