==> notebooks/03_baseline.ipynb <==
import pandas as pd
from sklearn.dummy import DummyRegressor
from sklearn.metrics import mean_absolute_error, mean_squared_error, r2_score
import numpy as np

# load feature engineered datasets
train_df = pd.read_csv('C:\\Users\\Mustapha\\Desktop\\ML-Project-Scratch\\data\\processed\\feature_engineered_train.csv')
eval_df = pd.read_csv('C:\\Users\\Mustapha\\Desktop\\ML-Project-Scratch\\data\\processed\\feature_engineered_eval.csv')

# ============================================
# 3. Define target & features
# ============================================
target = "price"
X_train = train_df.drop(columns=[target])
y_train = train_df[target]

X_eval = eval_df.drop(columns=[target])
y_eval = eval_df[target]

# ============================================
# 4. Baseline Model (Dummy Regressor)
# ============================================
baseline = DummyRegressor(strategy="median")   # median is better for skewed prices
baseline.fit(X_train, y_train)

# ============================================
# 5. Predictions & Metrics
# ============================================
y_pred = baseline.predict(X_eval)

mae = mean_absolute_error(y_eval, y_pred)
rmse = np.sqrt(mean_squared_error(y_eval, y_pred))
r2 = r2_score(y_eval, y_pred)

print("Baseline Dummy Regressor (median)")
print(f"MAE:  {mae:,.2f}")
print(f"RMSE: {rmse:,.2f}")
print(f"R²:   {r2:.4f}")
